refactor(mba): keep scalar metric in shared catalog
